refactor(ops): split 839-line db.py into 5 focused modules + unify seed data
- Create ops/_seed_data.py as single source of truth for blueprint seed data
- Split ops/db.py into: _schema (DDL), _seed (all seed data), _blueprints CRUD,
  _executions (record + stats), _dashboard (query helpers)
- __init__.py re-exports all public APIs for backward compat
- Remove duplicate SEED_VERSIONS/blueprints_data from ops/main.py
- Replace magic numbers (19480000, 4513, 4242) in _recalc_stats with constants
- Add _AgentResult dataclass and _execute_via() helper in executor.py to
  collapse the repeated if/elif/else dispatch across generate_plan/execute_step/reflect
