+++ apps/ops/db/_main.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+# apps/ops/db/_main.py — kept as import path anchor for existing consumers.
+# All logic moved to sub-modules. Import from here for backward compatibility:
+#   from apps.ops.db import get_db, init_db, ...
+# Or import directly from sub-modules:
+#   from apps.ops.db._schema      import get_db
+#   from apps.ops.db._blueprints import get_blueprint_config
+from apps.ops.db import (  # noqa: E402, F401
+    BASE_DIR,
+    DB_PATH,
+    OPENCLAW_BUILTIN_TOOLS,
+    get_activity_feed,
+    get_blueprint_config,
+    get_capability_dist,
+    get_cursor,
+    get_db,
+    get_recent_executions,
+    get_status_dist,
+    get_task_detail,
+    get_task_trend,
+    get_token_daily,
+    init_db,
+    record_execution,
+    run_schema,
+    save_blueprint_config,
+    seed_all,
+    seed_tools,
+)
